requirements.txt
python-telegram-bot == 12.7
requests == 2.23
